Update patch set 1
Patch Set 1:

Thanks for reviewing this. This patch was WIP and was not ready for review yet. I like to kind of work off gerrit, so to get the early input and share the code, sorry for the confusion this created.

Patch-set: 1
Reviewer: Gerrit User 3087 <3087@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0