enhance CadiFilter to support more noauthz
CadiFilter does hreq.getPathInfo() when getting the current path. This
can return null in the case of just a / in path or based on the
servlet-mapping section of pom (some engines will mark pathinfo as null
for url-pattern with a trailing star for instance). This change strips
out the context path and treats anything after it as what to compare
which makes it immune to changes in url-pattern.

Issue-ID: AAF-726
Change-Id: I7da63b347984dd816a6c24bf876c463a7ee71523
Signed-off-by: DR695H <dr695h@att.com>