Update patch set 1
Patch Set 1:

i can certainly tweak for any null pointers. 

As for why this is better, is it removes the need for knowing how the servlet path is defined. An example why is in https://javaee.github.io/servlet-spec/downloads/servlet-3.1/Final/servlet-3_1-final.pdf section 3.5: Servlet Path: The path section that directly corresponds to the mapping which activated this request. This path starts with a ’/’ character except in the case where the request is matched with the ‘/*’ or ““ pattern, in which case it is an empty string.

this means that a servlet path can be very variable, which doesnt match what you would expect often. My fix makes it so you always get the path after the context path, which more closely matches what you do with the tomcat valve and is more consistent.

Patch-set: 1