Update patch set 9
Patch Set 9:

There were a number of other issues, like not being able to see the "system.log", because no longer root, which causes failure to init, etc.

However, I'm solving these for the moment, and testing.  If I can get it to work, in Docker and in Helm,  I'll check it in.

Patch-set: 9