Fix CADI Connection to Remote DNS AAF servers on proxied network
Java CADI client code and web filter has been fixed to allow
specification of a fixed hostname for connecting to AAF
service and avoiding use of the AAF Locator service to bootstrap.
The CADI code was also updated to avoid pre-resolving DNS in this
scenario.  This fix is necessary to address use cases where the
AAF servers live in a different network domain from the CADI
client where the access point is through a proxied endpoint and
the DNS name is unresolvable on the network where the CADI client
code is executing.

Issue-ID: AAF-1123
Signed-off-by: Rene Glover <rene.glover@att.com>
Change-Id: I8835b7445e9bc1f6e89380462062748b53e078e5