Update patch set 3
Patch Set 3:

The biggest operational change I see is the change of the default of needing "/locate/" to indicate desire for Default Locator, versus lack of "/locate/" allows for use of (exiting, but improved SingleLocator".  There may need to be some testing to validate it works with all the existing Automated Services within ONAP, however, it is a reasonable change, consistent with the intent of "SingleLocator".

Patch-set: 3
Reviewer: Gerrit User 3289 <3289@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>