Update patch set 2
Patch Set 2:

Thanks for your contribution.

I think the expected way to correct a review request is to push again with the same Change-Id in the commit.  Then, the correction would be combined with the original review request.  This is best because then reviewers would accept only one review request.

I don't have objection to this change going in.

Patch-set: 2
CC: Gerrit User 6031 <6031@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>