Update patch set 1
Patch Set 1:

Well, I don`t need this immediately, as I am not completely blocked, but it would be nice to release a container after this other high-prio fix testing is completed. I`ll be looking at aaf/authz images.
(BTW. I noticed we have here a number of new fields filled in in the role configuration. 
This might be worth to test - after an image is available).

Patch-set: 1