Update patch set 1
Patch Set 1: Code-Review-1

Thanks Rupinderjeet,

I expect a unit test to check the value of a variant, or expect an exception.  These unit tests don't do either.  Fixing the code smell hides this deeper problem with the tests.


For example, in many tests of this change, there is already an assert, but commented out.  They check a variant condition.  These asserts should not be commented out.  Someone commented them out to avoid a test fail?  If so, the test itself is no good anymore or the implementation is broken.  Which?

The tests in the logobject should be checking the output of the log request, or merely succeed because there is no exception.  I'd like you to look into doing either.  If you chose the latter, please do something appropriate in junit assert to demonstrate that check.

Patch-set: 1
Reviewer: Gerrit User 6031 <6031@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1